lam support.

+++ H/Yapproto.h
@@ -389,7 +389,7 @@ void	STD_PROTO(Yap_InitUserBacks,(void));
 /* utilpreds.c */
 Term	STD_PROTO(Yap_CopyTerm,(Term));
 int	STD_PROTO(Yap_Variant,(Term, Term));
-int	STD_PROTO(Yap_ExportTerm,(Term, char *, size_t));
+size_t	STD_PROTO(Yap_ExportTerm,(Term, char *, size_t));
 Term	STD_PROTO(Yap_ImportTerm,(char *));
 int	STD_PROTO(Yap_IsListTerm,(Term));
 Term	STD_PROTO(Yap_CopyTermNoShare,(Term));
